feat(系统设置): 建某人的一揽子优化
